backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
executing==2.2.1
frozenlist==1.8.0
genai==2.1.0
google==3.0.0
google-auth==2.43.0
google-genai==1.52.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.2.1
multidict==6.7.0
openai==0.27.10
parso==0.8.5
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.3.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0